Cambio id a long en mapping file

+++ PliP/src/com/plip/persistence/dao/impls/PlipRoleDaoImpl.java
@@ -38,7 +38,7 @@ public class PlipRoleDaoImpl implements PlipRoleDao {
 	}
 
 	@Override
-	public PlipRole getRole(int idRole) {
+	public PlipRole getRole(Long idRole) {
 		SessionFactory factory = DaoManager.createSessionFactory();
 		Session session = factory.openSession();
 		Transaction tx = null;
